Fixes found by @mrandtx (Marc Randolph): Just got back to looking at this a bit more.
Not sure what you need it for (I haven't read up on the RPi),  but the SW1 botton won't work with the 10k in series.  Too weak against the 1k.  You could make it somewhere between 0 and 100 Ohms and it should be fine.
Also I don't know if there is ever a case for TestEn to be anything except pulled to ground - but the connection to GND on this schematic isn't making contact.  And I believe you already mentioned the 2.0 USB polarity.